+++ src/main/java/coupon/exception/coupon/CouponDiscountRateException.java
@@ -1,4 +1,4 @@
-package coupon.exception;
+package coupon.exception.coupon;
 
 public class CouponDiscountRateException extends CouponException {
