+++ src/main/java/coupon/exception/coupon/CouponException.java
@@ -1,4 +1,4 @@
-package coupon.exception;
+package coupon.exception.coupon;
 
 
 public class CouponException extends RuntimeException {
