+++ src/main/java/coupon/exception/coupon/CouponNotFoundException.java
@@ -1,4 +1,4 @@
-package coupon.exception;
+package coupon.exception.coupon;
 
 public class CouponNotFoundException extends CouponException {
