+++ src/main/java/coupon/repository/coupon/CouponRepository.java
@@ -1,6 +1,6 @@
-package coupon.repository;
+package coupon.repository.coupon;
 
-import coupon.entity.Coupon;
+import coupon.entity.coupon.Coupon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CouponRepository extends JpaRepository<Coupon, Long> {
