+++ src/test/java/coupon/service/coupon/CouponServiceTest.java
@@ -1,11 +1,11 @@
-package coupon.service;
+package coupon.service.coupon;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import coupon.entity.Coupon;
-import coupon.entity.CouponCategory;
-import coupon.exception.CouponNotFoundException;
+import coupon.entity.coupon.Coupon;
+import coupon.entity.coupon.CouponCategory;
+import coupon.exception.coupon.CouponNotFoundException;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
